requirements.txt
flask
Pillow
numpy
scipy
gunicorn
